hook refactor
